+++ src/Symfony/Component/EventDispatcher/Debug/TraceableEventDispatcher.php
@@ -32,6 +32,7 @@ class TraceableEventDispatcher implements TraceableEventDispatcherInterface
     private $called;
     private $dispatcher;
     private $wrappedListeners;
+    private $orphanedEvents;
 
     public function __construct(EventDispatcherInterface $dispatcher, Stopwatch $stopwatch, LoggerInterface $logger = null)
     {
@@ -40,6 +41,7 @@ class TraceableEventDispatcher implements TraceableEventDispatcherInterface
         $this->logger = $logger;
         $this->called = array();
         $this->wrappedListeners = array();
+        $this->orphanedEvents = array();
     }
 
     /**
@@ -247,6 +249,10 @@ class TraceableEventDispatcher implements TraceableEventDispatcherInterface
 
     private function preProcess($eventName)
     {
+        if (false === $this->dispatcher->hasListeners($eventName)) {
+            $this->orphanedEvents[] = $eventName;
+        }
+
         foreach ($this->dispatcher->getListeners($eventName) as $listener) {
             $priority = $this->getListenerPriority($eventName, $listener);
             $wrappedListener = new WrappedListener($listener, null, $this->stopwatch, $this);
@@ -318,5 +324,15 @@ class TraceableEventDispatcher implements TraceableEventDispatcherInterface
         }
 
         return 1;
+    }
+
+    /**
+     * Gets the orphaned events.
+     *
+     * @return array An array of orphaned events
+     */
+    public function getOrphanedEvents()
+    {
+        return $this->orphanedEvents;
     }
 }