+++ src/Symfony/Component/EventDispatcher/Tests/Debug/TraceableEventDispatcherTest.php
@@ -153,6 +153,40 @@ class TraceableEventDispatcherTest extends TestCase
         $this->assertCount(2, $dispatcher->getCalledListeners());
     }
 
+    public function testItReturnsNoOrphanedEventsWhenCreated()
+    {
+        // GIVEN
+        $tdispatcher = new TraceableEventDispatcher(new EventDispatcher(), new Stopwatch());
+        // WHEN
+        $events = $tdispatcher->getOrphanedEvents();
+        // THEN
+        $this->assertEmpty($events);
+    }
+
+    public function testItReturnsOrphanedEventsAfterDispatch()
+    {
+        // GIVEN
+        $tdispatcher = new TraceableEventDispatcher(new EventDispatcher(), new Stopwatch());
+        $tdispatcher->dispatch('foo');
+        // WHEN
+        $events = $tdispatcher->getOrphanedEvents();
+        // THEN
+        $this->assertCount(1, $events);
+        $this->assertEquals(array('foo'), $events);
+    }
+
+    public function testItDoesNotReturnHandledEvents()
+    {
+        // GIVEN
+        $tdispatcher = new TraceableEventDispatcher(new EventDispatcher(), new Stopwatch());
+        $tdispatcher->addListener('foo', function () {});
+        $tdispatcher->dispatch('foo');
+        // WHEN
+        $events = $tdispatcher->getOrphanedEvents();
+        // THEN
+        $this->assertEmpty($events);
+    }
+
     public function testLogger()
     {
         $logger = $this->getMockBuilder('Psr\Log\LoggerInterface')->getMock();