+++ src/Symfony/Bundle/WebProfilerBundle/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 4.0.0
 -----
 
+ * added information about orphaned events
  * removed the `WebProfilerExtension::dumpValue()` method
  * removed the `getTemplates()` method of the `TemplateManager` class in favor of the ``getNames()`` method
  * removed the `web_profiler.position` config option and the